+++ CHANGELOG.md
@@ -21,6 +21,9 @@ FlatLaf Change Log
 - TabbedPane: Support closable tabs. (PR #193; issues #31 and #40)
 - TabbedPane: Support minimum or maximum tab widths. (set client property
   `JTabbedPane.minimumTabWidth` or `JTabbedPane.maximumTabWidth` to an integer)
+- TabbedPane: Support alignment of tab area. (set client property
+  `JTabbedPane.tabAreaAlignment` to `"leading"`, `"trailing"`, `"center"` or
+  `"fill"`)
 - Support painting separator line between window title and content (use UI value
   `TitlePane.borderColor`). (issue #184)
 - Extras: `FlatSVGIcon` now allows specifying icon width and height in